+++ app/server/api_read_test/testdata/product_test/TestUser_GetUserByID_With_Goldie/正常系.golden.json
