+++ app/server/api_write_test/testdata/product_test/TestProduct_PostProducts_With_Goldie/正常系.golden.json
